Move Routes to a component (#652)
* Move Routes to a component

* Move loginFormContainer route

* Move routes to the component

* Move Routes into containers folder
